[move] dto 클래스 이동
